// js/src/types/resources/Bundle.ts
 
 
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TIdentifier } from '../partials/Identifier';
import { TInstant } from '../simpleTypes/Instant';
import { TUnsignedInt } from '../simpleTypes/UnsignedInt';
import { TBundleLink } from '../partials/BundleLink';
import { TBundleEntry } from '../partials/BundleEntry';
import { TSignature } from '../partials/Signature';

export type TBundle = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  identifier?: TIdentifier;
  type: string;
  timestamp?: TInstant;
  total?: TUnsignedInt;
  link?: TBundleLink[];
  entry?: TBundleEntry[];
  signature?: TSignature;
};