// js/src/types/partials/BundleEntry.ts
 
 
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from './Extension';
import { TBundleLink } from './BundleLink';
import { TUri } from '../simpleTypes/Uri';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TBundleSearch } from './BundleSearch';
import { TBundleRequest } from './BundleRequest';
import { TBundleResponse } from './BundleResponse';

export type TBundleEntry = {
  id?: string;
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  link?: TBundleLink[];
  fullUrl?: TUri;
  resource?: TResourceContainer;
  search?: TBundleSearch;
  request?: TBundleRequest;
  response?: TBundleResponse;
};