// js/src/types/resources/Composition.ts
 
 
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TDateTime } from '../simpleTypes/DateTime';
import { TCompositionAttester } from '../partials/CompositionAttester';
import { TCompositionRelatesTo } from '../partials/CompositionRelatesTo';
import { TCompositionEvent } from '../partials/CompositionEvent';
import { TCompositionSection } from '../partials/CompositionSection';

export type TComposition = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier;
  status: string;
  type: TCodeableConcept;
  category?: TCodeableConcept[];
  subject?: TReference;
  encounter?: TReference;
  date: TDateTime;
  author: TReference[];
  title: string;
  confidentiality?: string;
  attester?: TCompositionAttester[];
  custodian?: TReference;
  relatesTo?: TCompositionRelatesTo[];
  event?: TCompositionEvent[];
  section?: TCompositionSection[];
};