// js/src/types/partials/CompositionSection.ts
 
 
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from './Extension';
import { TCodeableConcept } from './CodeableConcept';
import { TReference } from './Reference';
import { TNarrative } from './Narrative';

export type TCompositionSection = {
  id?: string;
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  title?: string;
  code?: TCodeableConcept;
  author?: TReference[];
  focus?: TReference;
  text?: TNarrative;
  mode?: string;
  orderedBy?: TCodeableConcept;
  entry?: TReference[];
  emptyReason?: TCodeableConcept;
  section?: TCompositionSection[];
};